Initial commit
This commit adds my existing images (`.tex` files, with `tikz`
pictures), as well as two scripts: one for cleaning all garbage, and one
for generating all images. Perhaps in the future I would want to work
with a `template.tex` or similar file, as to not repeat a few lines in
each `.tex` file. Another further improvement is an option to build only
1 image, as opposed to a script that builds all.